requirements.txt
opencv-python==4.10.0.84
tensorboardX
numpy==1.24.0
nibabel
omegaconf
matplotlib
monai==0.5.3
segmentation_models_pytorch==0.4.0
SimpleITK==2.2.1
tqdm==4.67.1
scikit-image==0.20.0
scipy==1.9.1
